fix(android): fix crash after login - DTO defaults + race condition
- Add default values to all DTO fields (DashboardDto, WebSocketDto)
  to prevent MissingFieldException when backend omits optional fields
- Fix bandwidth_limit_mbps type Int? → Float? in ProfileResponseDto
  and DeviceResponseDto to match backend float values
- Fix WebSocketManager network callback race condition that could
  clear activeUrl during dashboard API requests
- Add global crash handler in TonbilApp for debugging

+++ android/app/src/main/java/com/tonbil/aifirewall/data/remote/dto/DeviceDto.kt
@@ -17,7 +17,7 @@ data class DeviceResponseDto(
     @SerialName("last_seen") val lastSeen: String? = null,
     @SerialName("total_online_seconds") val totalOnlineSeconds: Int = 0,
     @SerialName("last_online_start") val lastOnlineStart: String? = null,
-    @SerialName("bandwidth_limit_mbps") val bandwidthLimitMbps: Int? = null,
+    @SerialName("bandwidth_limit_mbps") val bandwidthLimitMbps: Float? = null,
 )
 
 @Serializable
